+++ Assets/Scripts/Weapons/StartingWeapon.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class StartingWeapon : BaseWeapon
 {
 
-    public override void Fire(Vector3 a_direction, float damagePerProjectile)
+    public override void Fire(Vector3 a_direction, int damagePerProjectile)
     {
 
         base.PoolToActive(a_direction, damagePerProjectile, 1);
